Benchmark Recipe Modification

+++ SLIDE-x-BENCH/RECIPE/bs/scnd.c
@@ -4,7 +4,7 @@
 #include <8051.h>
 #include <values.h>
 
-typedef float TARGET_TYPE;
+typedef int64_t TARGET_TYPE;
 typedef uint8_t TARGET_INDEX;
 
 //void prototype(int8_t n, int8_t arr[n], int8_t key);
