+++ SLIDE-x-BENCH/RECIPE/fdct/frst.c
@@ -12,7 +12,7 @@
 # endif
 #endif
 
-typedef float TARGET_TYPE;
+typedef double TARGET_TYPE;
 typedef uint8_t TARGET_INDEX;
 
 TARGET_INDEX i, j, k;
