+++ SLIDE-x-BENCH/RECIPE/matrix_mult/frst.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include <values.h>
 
-typedef float TARGET_TYPE;
+typedef double TARGET_TYPE;
 typedef uint8_t TARGET_INDEX;
 
 #ifndef RES
